Solutions to day 8 and attempts at day 9. Deleted day 7 because no longer compatible with incode computer

+++ src/main/java/computer/io/ComputerOutput.java
@@ -2,5 +2,5 @@ package computer.io;
 
 public interface ComputerOutput {
 
-    void output(int value);
+    void output(long value);
 }
