+++ src/main/java/computer/operations/Operation.java
@@ -1,11 +1,10 @@
 package computer.operations;
 
 import java.util.List;
-import java.util.Queue;
 
 public interface Operation {
 
-    int perform(List<Integer> program, int position);
+    int perform(List<Long> program, int position);
 
     int opCode();
